BEN-42 BEN-45 #resolve

+++ app/controllers/followings_controller.rb
@@ -18,20 +18,17 @@ class FollowingsController < ApplicationController
   def accept_following
     user = User.find_by_id(params[:id])
     follower = User.find_by_id(params[:follower_id])
-    following_request = FollowingRequest.where(user_id: follower.id, followee_id: user.id)
+    following_request = FollowingRequest.where(user_id: follower.id, followee_id: user.id).take!
     return not_authorized unless current_user == user
     
-    if following_request.exists?
-      create(follower.id, user.id, following_request)
-    else
-      record_not_found
-    end
+    create(follower.id, user.id, following_request)
   end
 
   def destroy
-    return not_authorized unless confirm_owner(params[:id])
+    following = Following.find_by_user_id_and_followee_id!(params[:id], params[:followee_id])
+    return not_authorized unless current_user == following.user
     
-    if Following.find_by_user_id_and_followee_id(params[:id], params[:followee_id]).destroy
+    if following.destroy    
       resource_destroyed      
     else
       # TODO: Log this?
@@ -65,9 +62,7 @@ class FollowingsController < ApplicationController
   def create(user_id, followee_id, following_request=nil)
     if Following.find_or_create_by(user_id: user_id, followee_id: followee_id) 
 
-      if following_request == FollowingRequest.new
-        following_request.destroy
-      end 
+      following_request.destroy unless following_request.nil?
 
       resource_created        
     else
